refactor: move snapshot tests for the gen command
