+++ Cargo.toml
@@ -272,6 +272,11 @@ name = "execution_spec"
 harness = false
 
 [[test]]
+name = "cli_spec"
+path = "tests/cli/gen.rs"
+harness = false
+
+[[test]]
 name = "json_to_config_spec"
 path = "src/core/generator/tests/json_to_config_spec.rs"
 harness = false
