+++ tests/cli/fixtures/generator/deezer.json
