+++ tests/cli/fixtures/generator/json_proto_mix_config.json
