+++ tests/cli/fixtures/generator/jsonplaceholder.json
